+++ src/textual/layouts/horizontal.py
@@ -101,6 +101,7 @@ class HorizontalLayout(Layout):
                 (next_x - x.__floor__()).__floor__(),
                 content_height.__floor__(),
             )
+            absolute = styles.has_rule("position") and styles.position == "absolute"
             add_placement(
                 _WidgetPlacement(
                     region,
@@ -110,10 +111,10 @@ class HorizontalLayout(Layout):
                     0,
                     False,
                     overlay,
-                    styles.has_rule("position") and styles.position == "absolute",
+                    absolute,
                 )
             )
-            if not overlay:
+            if not overlay and not absolute:
                 x = next_x + margin
 
         return placements